+++ build.js
@@ -5,38 +5,105 @@ const Color = require('tinycolor2');
 function getStyleDictionaryConfig(platform) {
   return {
     "source": [
-      "src/properties/globals/**/*.json",
-      `src/properties/platforms/${platform}/**/*.json`
+      "properties/globals/**/*.json",
+      `properties/platforms/${platform}/**/*.json`
     ],
     "platforms": {
       "web": {
         "transformGroup": "custom/web",
         "buildPath": `build/web/`,
         "files": [{
-          "destination": "_niten-tokens.scss",
-          "format": "scss/variables"
+          "destination": "design-tokens.scss",
+          "format": "scss/variables",
+          "options": {
+            "outputReferences": true
+          }
+        },{
+          "destination": "design-tokens.css",
+          "format": "css/variables",
+          "options": {
+          "outputReferences": true
+        }
         }]
       },
       "ios": {
         "transformGroup": "custom/ios",
         "buildPath": "build/ios/",
         "files": [{
-          "destination": "NitenTokens.swift",
+          "destination": "DesignTokensColors.swift",
+          "format": "ios-swift/enum.swift",
+          "className": "DesignTokensColors",
+          "filter": {
+            "attributes": {
+              "category": "color"
+            }
+          }
+        },{
+          "destination": "DesignTokensBorder.swift",
+          "format": "ios-swift/enum.swift",
+          "className": "DesignTokensBorder",
+          "type": "float",
+          "filter": {
+            "attributes": {
+              "category": "size",
+              "type": "border"
+            }
+          }
+        },{
+          "destination": "DesignTokensFontSize.swift",
+          "format": "ios-swift/enum.swift",
+          "className": "DesignTokensFontSize",
+          "type": "float",
+          "filter": {
+            "attributes": {
+              "category": "size",
+              "type": "font"
+            }
+          }
+        },{
+          "destination": "DesignTokensSpacings.swift",
+          "format": "ios-swift/enum.swift",
+          "className": "DesignTokensSpacings",
+          "type": "float",
+          "filter": {
+            "attributes": {
+              "category": "size",
+              "type": "global"
+            }
+          }
+        },{
+          "destination": "DesignTokensFont.swift",
+          "format": "ios-swift/enum.swift",
+          "className": "DesignTokensFont",
+          "type": "float",
+          "filter": {
+            "attributes": {
+              "category": "font"
+            }
+          }
+        },{
+          "destination": "DesignTokensShadows.swift",
           "format": "ios-swift/class.swift",
-          "className": "NitenTokens",
+          "className": "DesignTokensShadows",
+          "filter": {
+            "attributes": {
+              "category": "shadow"
+            }
+          }
         }]
       },
       "android": {
         "transformGroup": "custom/android",
+        "basePxFontSize": 14,
         "buildPath": "build/android/",
         "files": [{
-          "destination": "niten_colors.xml",
+          "destination": "tokens_colors.xml",
           "format": "android/colors"
         },{
-          "destination": "niten_font_dimens.xml",
+          "destination": "tokens_font_dimens.xml",
           "format": "android/fontDimens"
         },{
-          "destination": "niten_dimens.xml",
+          "destination": "tokens_dimens.xml",
           "format": "android/dimens"
         }]
       },
@@ -49,6 +116,10 @@ function isOffset(prop) {
 }
 
 function isSize(prop) {
+  return prop.attributes.category === 'size';
+}
+
+function isFontSize (prop) {
   return prop.attributes.category === 'size';
 }
 
@@ -74,10 +145,11 @@ StyleDictionaryPackage.registerTransform({
 });
 
 StyleDictionaryPackage.registerTransform({
-  name: 'name/ti/camel',
-  type: 'name',
-  transformer: function(prop, options) {
-    return _.camelCase([options.prefix].concat(prop.path).join(' '));
+  name: 'size/rem',
+  type: 'value',
+  matcher: isFontSize,
+  transformer: function (prop) {
+    return `rem(${(parseFloat(prop.value, 10)).toFixed(2)})`;
   }
 });
 
@@ -91,7 +163,7 @@ StyleDictionaryPackage.registerTransform({
 });
 
 StyleDictionaryPackage.registerTransform({
-  name: 'color/UIColorSwift',
+  name: 'color/DesignSystemColor',
   type: 'value',
   matcher: isColor,
   transformer: function(prop) {
@@ -99,7 +171,7 @@ StyleDictionaryPackage.registerTransform({
     const rFixed = (r / 255.0).toFixed(2);
     const gFixed = (g / 255.0).toFixed(2);
     const bFixed = (b / 255.0).toFixed(2);
-    return `UIColor(red: ${rFixed}, green: ${gFixed}, blue: ${bFixed}, alpha: ${a})`;
+    return `DesignSystemColor(red: ${rFixed}, green: ${gFixed}, blue: ${bFixed}, alpha: ${a})`;
   }
 });
 
@@ -109,6 +181,15 @@ StyleDictionaryPackage.registerTransform({
   matcher: isShadowRadius,
   transformer: function(prop) {
     return `CGFloat(${(parseFloat(prop.value, 10)).toFixed(2)})`;
+  }
+});
+
+StyleDictionaryPackage.registerTransform({
+  name: 'size/compose/remToDp',
+  type: 'value',
+  matcher: isSize,
+  transformer: function(prop) {
+    return parseFloat(prop.value, 10) + 'dp';
   }
 });
 
@@ -137,17 +218,17 @@ StyleDictionaryPackage.registerTransform({
 
 StyleDictionaryPackage.registerTransformGroup({
   name: 'custom/web',
-  transforms: ["attribute/cti", "name/cti/kebab", "color/sixDigitsToThreeDigits"]
+  transforms: ["attribute/cti", "name/cti/kebab", "color/css", "size/px"]
 });
 
 StyleDictionaryPackage.registerTransformGroup({
   name: 'custom/ios',
-  transforms: ["attribute/cti", "name/ti/camel", "color/UIColorSwift", "size/CGSize", "size/pxToCGFloat", "size/shadowRadiusToCGFloat"]
+  transforms: ["attribute/cti", "name/ti/camel", "color/DesignSystemColor", "size/CGSize", "size/pxToCGFloat", "size/shadowRadiusToCGFloat"]
 });
 
 StyleDictionaryPackage.registerTransformGroup({
   name: 'custom/android',
-  transforms: ["attribute/cti", "name/cti/snake", "color/hex", "size/dp"]
+  transforms: ["attribute/cti", "name/cti/snake", "color/hex", "size/dp", "size/sp"]
 });
 
 ['web', 'ios', 'android'].map(function (platform) {