+++ CHANGELOG.md
@@ -15,3 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added config to web platform using [style-dictionary](https://github.com/amzn/style-dictionary) (https://github.com/getninjas/niten-tokens/pull/5)
 
 [unreleased]: https://github.com/getninjas/niten-tokens/compare/v1.0.1-0...HEAD
+
+## 2.0.0 - 2021-08-06
+
+- Changed folder structure to present new build files
+- Added new design tokens with new values and formats
+- Removed Niten Tokens
+